Rename helper file to colWidthHelper
