+++ src/components/index.ts
@@ -11,4 +11,4 @@ export { SEO } from "./seo/seo"
 export { SubHeading } from "./subHeading/subHeading"
 export { RightPage } from "./rightPage/rightPage"
 export { TopHeader } from "./topHeader/topHeader"
-export { BoxWidthConversion } from "./bookSection/helper/formattingHelper"
+export { BoxWidthConversion } from "./bookSection/helper/colWidthHelper"
